Refactoring Page Components
-- Changed tabbed components to route tabs.
-- Code cleanup from various page componets and MainLayout
-- Move ProgressCircle Logic to Vuex for central management
-- Cleanup dark mode functionality in various pages.
-- Added names to all routes for future ease of management
