+++ src/boot/mixins.js
@@ -1,0 +1,7 @@
+import Vue from 'vue'
+import globalMixin from 'src/mixins/global'
+
+// "async" is optional
+export default async () => {
+  Vue.mixin(globalMixin)
+}
