--- src/css/app.sass
@@ -1,1 +1,0 @@
-// app global css in Sass form
