+++ src/css/app.scss
@@ -1,0 +1,48 @@
+@import "quasar.variables";
+@import "~quasar/src/css/variables.sass";
+
+
+body, .site-bg {
+	background: $grey-2;
+}
+.bg-shade{
+	background: #ffffff;
+}
+
+body.body--dark {
+	background: $blue-grey-9;
+
+	.q-card{
+		background: $blue-grey-9;
+	}
+
+	&.site-bg, .site-bg{
+		background: $blue-grey-9;
+	}
+
+	.bg-shade{
+		background: $blue-grey-7;
+	}
+}
+
+
+
+
+.border-1{
+	border: 1px solid;
+}
+
+.border-rounded-5{
+	border-radius: 5px;
+}
+
+.border-white{
+	border-color: white;
+}
+
+.border-black{
+	border-color: black;
+}
+.q-card{
+	background:$grey-4;
+}
