+++ src/mixins/global.js
@@ -1,0 +1,10 @@
+export default {
+  methods: {
+    showProgressSpinner () {
+      this.$store.commit('ui/showProgressSpinner')
+    },
+    hideProgressSpinner () {
+      this.$store.commit('ui/hideProgressSpinner')
+    }
+  }
+}
